requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
Pillow
python-decouple
requests
pytz
gunicorn
whitenoise
